refactor(ui): extract ChatPage starter-menu into a focused module
ChatPage.tsx was 1795 lines. Move the self-contained, pure (no-React)
starter-menu data + helpers — EXAMPLE_STARTERS, the StarterPrompt/cache
types, getTextList, and the localStorage-backed read/writeStarterMenuCache
— into ChatPage.starters.ts. ChatPage 1795 -> 1700; the starter logic is
now independently testable. Behaviour unchanged (pure move + import).

UI typecheck + lint + 347 tests + build green.
